Move client and server parts to separate folders
