+++ client/actions/detector.js
