+++ client/actions/devices.js
