+++ client/actions/media-stream.js
