+++ client/components/DetectorConfig.jsx
