+++ client/components/DeviceConnectionStatus.jsx
