+++ client/components/PushNotificationSwitch.jsx
