+++ client/containers/Device.jsx
