+++ client/containers/DeviceList.jsx
