+++ client/lib/detectors.js
