+++ client/lib/detectors/motion.js
