+++ client/lib/detectors/noise.js
