+++ client/lib/firebase.js
