+++ client/lib/install-sw-push-notifications.js
