+++ client/manifest.json
