+++ client/reducers/devices.js
