+++ client/reducers/incidents.js
