+++ client/reducers/peer.js
